backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pandas
requests